@@ -53,6 +53,7 @@ using MiNET.Net;
 using MiNET.Utils;
 using MiNET.Worlds;
 using Newtonsoft.Json;
+using Newtonsoft.Json.Converters;
 using Org.BouncyCastle.Crypto;
 using Org.BouncyCastle.Crypto.Agreement;
 using Org.BouncyCastle.Crypto.Parameters;
@@ -73,7 +74,8 @@ namespace MiNET.Client
 		public IPEndPoint ClientEndpoint { get; private set; }
 		public IPEndPoint ServerEndpoint { get; private set; }
 		private readonly DedicatedThreadPool _threadPool;
-		private short _mtuSize = 1192;
+		//private short _mtuSize = 1192;
+		private short _mtuSize = 1464;
 		private int _reliableMessageNumber = -1;
 		public Vector3 SpawnPoint { get; set; }
 		public long EntityId { get; set; }
@@ -103,8 +105,9 @@ namespace MiNET.Client
 
 		public MiNetClient(IPEndPoint endpoint, string username, DedicatedThreadPool threadPool)
 		{
-			Random random = new Random();
-			_clientGuid = 1111111 + random.Next() + random.Next();
+			byte[] buffer = new byte[8];
+			new Random().NextBytes(buffer);
+			_clientGuid = BitConverter.ToInt64(buffer, 0);
 
 			Username = username;
 			ClientId = new Random().Next();
@@ -330,7 +333,7 @@ namespace MiNET.Client
 
 			if (msgId <= (byte) DefaultMessageIdTypes.ID_USER_PACKET_ENUM)
 			{
-				DefaultMessageIdTypes msgIdType = (DefaultMessageIdTypes) msgId;
+				var msgIdType = (DefaultMessageIdTypes) msgId;
 
 				Packet message = PacketFactory.Create(msgId, receiveBytes, "raknet");
 
@@ -342,14 +345,14 @@ namespace MiNET.Client
 				{
 					case DefaultMessageIdTypes.ID_UNCONNECTED_PONG:
 					{
-						UnconnectedPong incoming = (UnconnectedPong) message;
+						var incoming = (UnconnectedPong) message;
 						OnUnconnectedPong(incoming, senderEndpoint);
 
 						break;
 					}
 					case DefaultMessageIdTypes.ID_OPEN_CONNECTION_REPLY_1:
 					{
-						OpenConnectionReply1 incoming = (OpenConnectionReply1) message;
+						var incoming = (OpenConnectionReply1) message;
 						if (incoming.mtuSize != _mtuSize) Log.Warn("Error, mtu differ from what we sent:" + incoming.mtuSize);
 						Log.Warn($"Server with ID {incoming.serverGuid} security={incoming.serverHasSecurity}");
 						_mtuSize = incoming.mtuSize;
@@ -547,8 +550,15 @@ namespace MiNET.Client
 
 		protected virtual void OnUnconnectedPong(UnconnectedPong packet, IPEndPoint senderEndpoint)
 		{
+			Log.Warn($"MOTD: {packet.serverName}");
+			Log.Warn($"from server {senderEndpoint}");
+
+			
 			if (!HaveServer)
 			{
+				string[] motdParts = packet.serverName.Split(';');
+				senderEndpoint.Port = int.Parse(motdParts[10]);
+				Log.Debug($"Connecting to {senderEndpoint}");
 				ServerEndpoint = senderEndpoint;
 				HaveServer = true;
 				SendOpenConnectionRequest1();
@@ -724,7 +734,7 @@ namespace MiNET.Client
 
 		private void OnNoFreeIncomingConnections(NoFreeIncomingConnections message)
 		{
-			Log.Error("No free connections from server ");
+			Log.Error($"No free connections from server {message.serverGuid}");
 			StopClient();
 		}
 
@@ -860,7 +870,7 @@ namespace MiNET.Client
 					}
 				}
 				{
-					ItemStacks slotData = new ItemStacks {recipe.Result};
+					ItemStacks slotData = new ItemStacks {recipe.Result.First()};
 					crafting.result = slotData;
 				}
 
@@ -1176,7 +1186,7 @@ namespace MiNET.Client
 
 		private int _numberOfChunks = 0;
 
-		public ConcurrentDictionary<Tuple<int, int>, bool> _chunks = new ConcurrentDictionary<Tuple<int, int>, bool>();
+		public ConcurrentDictionary<Tuple<int, int>, ChunkColumn> _chunks = new ConcurrentDictionary<Tuple<int, int>, ChunkColumn>();
 		public IndentedTextWriter _mobWriter;
 
 		public virtual void HandleBatch(McpeWrapper batch)
@@ -1333,8 +1343,11 @@ namespace MiNET.Client
 
 					Formatting = Formatting.Indented,
 				};
+				jsonSerializerSettings.Converters.Add(new StringEnumConverter());
 				jsonSerializerSettings.Converters.Add(new NbtIntConverter());
 				jsonSerializerSettings.Converters.Add(new NbtStringConverter());
+				jsonSerializerSettings.Converters.Add(new IPAddressConverter());
+				jsonSerializerSettings.Converters.Add(new IPEndPointConverter());
 
 				string result = JsonConvert.SerializeObject(message, jsonSerializerSettings);
 				Log.Debug($"> Receive: {message.Id} (0x{message.Id:x2}): {message.GetType().Name}\n{result}");
@@ -1366,7 +1379,33 @@ namespace MiNET.Client
 				return;
 			}
 
-			Log.DebugFormat("<    Send: {0}: {1} (0x{0:x2})", message.Id, message.GetType().Name);
+
+			if (verbosity == 0)
+			{
+				Log.Debug($"<    Send: {message.Id} (0x{message.Id:x2}): {message.GetType().Name}");
+			}
+			else if (verbosity == 1)
+			{
+				var jsonSerializerSettings = new JsonSerializerSettings
+				{
+					PreserveReferencesHandling = PreserveReferencesHandling.All,
+
+					Formatting = Formatting.Indented,
+				};
+				jsonSerializerSettings.Converters.Add(new StringEnumConverter());
+				jsonSerializerSettings.Converters.Add(new NbtIntConverter());
+				jsonSerializerSettings.Converters.Add(new NbtStringConverter());
+				jsonSerializerSettings.Converters.Add(new IPAddressConverter());
+				jsonSerializerSettings.Converters.Add(new IPEndPointConverter());
+
+				string result = JsonConvert.SerializeObject(message, jsonSerializerSettings);
+				Log.Debug($"<    Send: {message.Id} (0x{message.Id:x2}): {message.GetType().Name}\n{result}");
+			}
+			else if (verbosity == 2)
+			{
+				Log.Debug($"<    Send: {message.Id} (0x{message.Id:x2}): {message.GetType().Name}\n{Packet.HexDump(message.Bytes)}");
+			}
+
 		}
 
 		public void SendUnconnectedPing()
@@ -1374,6 +1413,7 @@ namespace MiNET.Client
 			var packet = new UnconnectedPing
 			{
 				pingId = Stopwatch.GetTimestamp() /*incoming.pingId*/,
+				guid = _clientGuid
 			};
 
 			var data = packet.Encode();
@@ -1427,6 +1467,7 @@ namespace MiNET.Client
 			byte[] data2 = new byte[_mtuSize - data.Length - 10];
 			Buffer.BlockCopy(data, 0, data2, 0, data.Length);
 
+			Log.Debug($"data lenght={data2.Length}");
 			SendData(data2);
 		}
 
