@@ -3,10 +3,10 @@
 // The contents of this file are subject to the Common Public Attribution
 // License Version 1.0. (the "License"); you may not use this file except in
 // compliance with the License. You may obtain a copy of the License at
-// https://github.com/NiclasOlofsson/MiNET/blob/master/LICENSE. 
-// The License is based on the Mozilla Public License Version 1.1, but Sections 14 
-// and 15 have been added to cover use of software over a computer network and 
-// provide for limited attribution for the Original Developer. In addition, Exhibit A has 
+// https://github.com/NiclasOlofsson/MiNET/blob/master/LICENSE.
+// The License is based on the Mozilla Public License Version 1.1, but Sections 14
+// and 15 have been added to cover use of software over a computer network and
+// provide for limited attribution for the Original Developer. In addition, Exhibit A has
 // been modified to be consistent with Exhibit B.
 // 
 // Software distributed under the License is distributed on an "AS IS" basis,
@@ -18,13 +18,12 @@
 // The Original Developer is the Initial Developer.  The Initial Developer of
 // the Original Code is Niclas Olofsson.
 // 
-// All portions of the code written by Niclas Olofsson are Copyright (c) 2014-2018 Niclas Olofsson. 
+// All portions of the code written by Niclas Olofsson are Copyright (c) 2014-2020 Niclas Olofsson.
 // All Rights Reserved.
 
 #endregion
 
 using System.Numerics;
-using log4net;
 using MiNET.Entities.World;
 using MiNET.Items;
 using MiNET.Utils;
@@ -34,9 +33,7 @@ namespace MiNET.Blocks
 {
 	public class Sand : Block
 	{
-		private static readonly ILog Log = LogManager.GetLogger(typeof(Sand));
-
-		private int _tickRate = 10;
+		private int _tickRate = 1;
 
 		public Sand() : base(12)
 		{
@@ -66,16 +63,12 @@ namespace MiNET.Blocks
 
 			if (!level.GetBlock(Coordinates + Level.Down).IsSolid)
 			{
-				//level.SetAir(Coordinates, false);
-				level.SetBlock(new Air() {Coordinates = Coordinates}, false, false);
+				level.SetAir(Coordinates);
 
 				var bbox = GetBoundingBox();
 				var d = (bbox.Max - bbox.Min) / 2;
 
-				new FallingBlock(level, this)
-				{
-					KnownPosition = new PlayerLocation(Coordinates.X + d.X, Coordinates.Y - 0.03f, Coordinates.Z + d.Z)
-				}.SpawnEntity();
+				new FallingBlock(level, this) {KnownPosition = new PlayerLocation(Coordinates.X + d.X, Coordinates.Y - 0.03f, Coordinates.Z + d.Z)}.SpawnEntity();
 			}
 		}
 
