@@ -346,7 +346,7 @@ namespace MiNET
 							else if (validationKey != null && validationKey.Equals(x5u, StringComparison.InvariantCultureIgnoreCase))
 							{
 								//TODO: Remove. Just there to be able to join with same XBL multiple times without crashing the server.
-								data.ExtraData.Identity = Guid.NewGuid().ToString();
+								//data.ExtraData.Identity = Guid.NewGuid().ToString();
 								_playerInfo.CertificateData = data;
 							}
 							else
@@ -361,7 +361,7 @@ namespace MiNET
 								}
 
 								//TODO: Remove. Just there to be able to join with same XBL multiple times without crashing the server.
-								data.ExtraData.Identity = Guid.NewGuid().ToString();
+								//data.ExtraData.Identity = Guid.NewGuid().ToString();
 								_playerInfo.CertificateData = data;
 							}
 						}
@@ -386,9 +386,6 @@ namespace MiNET
 							UseEncryption = Config.GetProperty("UseEncryptionForAll", false) || (Config.GetProperty("UseEncryption", true) && !string.IsNullOrWhiteSpace(_playerInfo.CertificateData.ExtraData.Xuid)),
 						};
 
-#if LINUX
-						_session.CryptoContext.UseEncryption = false;
-#else
 						if (_session.CryptoContext.UseEncryption)
 						{
 							// Use bouncy to parse the DER key
@@ -473,7 +470,6 @@ namespace MiNET
 
 							if (Log.IsDebugEnabled) Log.Warn($"Encryption enabled for {_session.Username}");
 						}
-#endif
 					}
 				}
 				if (!_session.CryptoContext.UseEncryption)
@@ -535,7 +531,15 @@ namespace MiNET
 		{
 		}
 
+		public void HandleMcpeTickSync(McpeTickSync message)
+		{
+		}
+
 		public void HandleMcpeLevelSoundEvent(McpeLevelSoundEvent message)
+		{
+		}
+
+		public void HandleMcpeClientCacheStatus(McpeClientCacheStatus message)
 		{
 		}
 
