@@ -24,6 +24,7 @@
 #endregion
 
 using System;
+using System.Collections.Generic;
 using System.Diagnostics;
 using System.IO;
 using System.Linq;
@@ -574,7 +575,7 @@ namespace MiNET.Net
 				var parts = endpoint.Address.ToString().Split('.');
 				foreach (var part in parts)
 				{
-					Write((byte) byte.Parse(part));
+					Write((byte) ~byte.Parse(part));
 				}
 				Write((short) endpoint.Port, true);
 			}
@@ -603,7 +604,7 @@ namespace MiNET.Net
 
 			if (ipVersion == 4)
 			{
-				string ipAddress = $"{ReadByte()}.{ReadByte()}.{ReadByte()}.{ReadByte()}";
+				string ipAddress = $"{(byte) ~ReadByte()}.{(byte) ~ReadByte()}.{(byte) ~ReadByte()}.{(byte) ~ReadByte()}";
 				address = IPAddress.Parse(ipAddress);
 				port = (ushort) ReadShort(true);
 			}
@@ -751,32 +752,50 @@ namespace MiNET.Net
 			return metadata;
 		}
 
-		public void Write(Transaction trans)
+		public void Write(Transaction transaction)
 		{
-			WriteVarInt((int) trans.TransactionType);
-			WriteUnsignedVarInt((uint) trans.Transactions.Count);
-			foreach (var record in trans.Transactions)
+			switch (transaction)
 			{
-				if (record is ContainerTransactionRecord)
+				case InventoryMismatchTransaction _:
+					WriteVarInt((int) McpeInventoryTransaction.TransactionType.InventoryMismatch);
+					break;
+				case ItemReleaseTransaction _:
+					WriteVarInt((int) McpeInventoryTransaction.TransactionType.ItemRelease);
+					break;
+				case ItemUseOnEntityTransaction _:
+					WriteVarInt((int) McpeInventoryTransaction.TransactionType.ItemUseOnEntity);
+					break;
+				case ItemUseTransaction _:
+					WriteVarInt((int) McpeInventoryTransaction.TransactionType.ItemUse);
+					break;
+				case NormalTransaction _:
+					WriteVarInt((int) McpeInventoryTransaction.TransactionType.Normal);
+					break;
+			}
+
+			WriteUnsignedVarInt((uint) transaction.TransactionRecords.Count);
+			foreach (var record in transaction.TransactionRecords)
+			{
+				switch (record)
 				{
-					var r = record as ContainerTransactionRecord;
-					WriteSignedVarInt(r.InventoryId);
-				}
-				else if (record is GlobalTransactionRecord)
-				{
-				}
-				else if (record is WorldInteractionTransactionRecord)
-				{
-					var r = record as WorldInteractionTransactionRecord;
-					WriteVarInt(r.Flags);
-				}
-				else if (record is CreativeTransactionRecord)
-				{
-				}
-				else if (record is CraftTransactionRecord)
-				{
-					var r = record as CraftTransactionRecord;
-					WriteVarInt(r.Action);
+					case ContainerTransactionRecord r:
+						WriteVarInt((int) McpeInventoryTransaction.InventorySourceType.Container);
+						WriteSignedVarInt(r.InventoryId);
+						break;
+					case GlobalTransactionRecord _:
+						WriteVarInt((int) McpeInventoryTransaction.InventorySourceType.Global);
+						break;
+					case WorldInteractionTransactionRecord r:
+						WriteVarInt((int) McpeInventoryTransaction.InventorySourceType.WorldInteraction);
+						WriteVarInt(r.Flags);
+						break;
+					case CreativeTransactionRecord _:
+						WriteVarInt((int) McpeInventoryTransaction.InventorySourceType.Creative);
+						break;
+					case CraftTransactionRecord r:
+						WriteVarInt((int) McpeInventoryTransaction.InventorySourceType.Crafting);
+						WriteVarInt((int) r.Action);
+						break;
 				}
 
 				WriteVarInt(record.Slot);
@@ -784,34 +803,34 @@ namespace MiNET.Net
 				Write(record.NewItem);
 			}
 
-			switch (trans.TransactionType)
+			switch (transaction)
 			{
-				case McpeInventoryTransaction.TransactionType.Normal:
-				case McpeInventoryTransaction.TransactionType.InventoryMismatch:
+				case NormalTransaction _:
+				case InventoryMismatchTransaction _:
 					break;
-				case McpeInventoryTransaction.TransactionType.ItemUse:
-					WriteVarInt(trans.ActionType);
-					Write(trans.Position);
-					WriteSignedVarInt(trans.Face);
-					WriteSignedVarInt(trans.Slot);
-					Write(trans.Item);
-					Write(trans.FromPosition);
-					Write(trans.ClickPosition);
-					WriteUnsignedVarInt(trans.BlockRuntimeId);
+				case ItemUseTransaction t:
+					WriteVarInt((int)t.ActionType);
+					Write(t.Position);
+					WriteSignedVarInt(t.Face);
+					WriteSignedVarInt(t.Slot);
+					Write(t.Item);
+					Write(t.FromPosition);
+					Write(t.ClickPosition);
+					WriteUnsignedVarInt(t.BlockRuntimeId);
 					break;
-				case McpeInventoryTransaction.TransactionType.ItemUseOnEntity:
-					WriteVarLong(trans.EntityId);
-					WriteVarInt(trans.ActionType);
-					WriteSignedVarInt(trans.Slot);
-					Write(trans.Item);
-					Write(trans.FromPosition);
-					Write(trans.ClickPosition);
+				case ItemUseOnEntityTransaction t:
+					WriteVarLong(t.EntityId);
+					WriteVarInt((int) t.ActionType);
+					WriteSignedVarInt(t.Slot);
+					Write(t.Item);
+					Write(t.FromPosition);
+					Write(t.ClickPosition);
 					break;
-				case McpeInventoryTransaction.TransactionType.ItemRelease:
-					WriteVarInt(trans.ActionType);
-					WriteSignedVarInt(trans.Slot);
-					Write(trans.Item);
-					Write(trans.FromPosition);
+				case ItemReleaseTransaction t:
+					WriteVarInt((int) t.ActionType);
+					WriteSignedVarInt(t.Slot);
+					Write(t.Item);
+					Write(t.FromPosition);
 					break;
 				default:
 					break;
@@ -820,41 +839,33 @@ namespace MiNET.Net
 
 		public Transaction ReadTransaction()
 		{
-			var trans = new Transaction();
-
-			trans.TransactionType = (McpeInventoryTransaction.TransactionType) ReadVarInt();
-
+			var transactions = new List<TransactionRecord>();
+			var transactionType = (McpeInventoryTransaction.TransactionType) ReadVarInt();
 			var count = ReadUnsignedVarInt();
 			for (int i = 0; i < count; i++)
 			{
-				TransactionRecord record = null;
+				TransactionRecord record;
 				int sourceType = ReadVarInt();
 				switch ((McpeInventoryTransaction.InventorySourceType) sourceType)
 				{
 					case McpeInventoryTransaction.InventorySourceType.Container:
 						record = new ContainerTransactionRecord()
 						{
-							Source = sourceType,
 							InventoryId = ReadSignedVarInt()
 						};
 						break;
 					case McpeInventoryTransaction.InventorySourceType.Global:
-						record = new GlobalTransactionRecord()
-						{
-							Source = sourceType,
-						};
+						record = new GlobalTransactionRecord();
 						break;
 					case McpeInventoryTransaction.InventorySourceType.WorldInteraction:
 						record = new WorldInteractionTransactionRecord()
 						{
-							Source = sourceType,
 							Flags = ReadVarInt()
 						};
 						break;
 					case McpeInventoryTransaction.InventorySourceType.Creative:
 						record = new CreativeTransactionRecord()
 						{
-							Source = sourceType,
 							InventoryId = 0x79
 						};
 						break;
@@ -862,8 +873,7 @@ namespace MiNET.Net
 					case McpeInventoryTransaction.InventorySourceType.Crafting:
 						record = new CraftTransactionRecord()
 						{
-							Source = sourceType,
-							Action = ReadSignedVarInt()
+							Action = (McpeInventoryTransaction.CraftingAction) ReadSignedVarInt()
 						};
 						break;
 					default:
@@ -874,43 +884,57 @@ namespace MiNET.Net
 				record.Slot = ReadVarInt();
 				record.OldItem = ReadItem();
 				record.NewItem = ReadItem();
-				trans.Transactions.Add(record);
+				transactions.Add(record);
 			}
 
-			switch (trans.TransactionType)
+			Transaction transaction = null;
+			switch (transactionType)
 			{
 				case McpeInventoryTransaction.TransactionType.Normal:
+					transaction = new NormalTransaction();
+					break;
 				case McpeInventoryTransaction.TransactionType.InventoryMismatch:
+					transaction = new InventoryMismatchTransaction();
 					break;
 				case McpeInventoryTransaction.TransactionType.ItemUse:
-					trans.ActionType = ReadVarInt();
-					trans.Position = ReadBlockCoordinates();
-					trans.Face = ReadSignedVarInt();
-					trans.Slot = ReadSignedVarInt();
-					trans.Item = ReadItem();
-					trans.FromPosition = ReadVector3();
-					trans.ClickPosition = ReadVector3();
-					trans.BlockRuntimeId = ReadUnsignedVarInt();
+					transaction = new ItemUseTransaction()
+					{
+						ActionType = (McpeInventoryTransaction.ItemUseAction) ReadVarInt(),
+						Position = ReadBlockCoordinates(),
+						Face = ReadSignedVarInt(),
+						Slot = ReadSignedVarInt(),
+						Item = ReadItem(),
+						FromPosition = ReadVector3(),
+						ClickPosition = ReadVector3(),
+						BlockRuntimeId = ReadUnsignedVarInt()
+					};
 					break;
 				case McpeInventoryTransaction.TransactionType.ItemUseOnEntity:
-					trans.EntityId = ReadVarLong();
-					trans.ActionType = ReadVarInt();
-					trans.Slot = ReadSignedVarInt();
-					trans.Item = ReadItem();
-					trans.FromPosition = ReadVector3();
-					trans.ClickPosition = ReadVector3();
+					transaction = new ItemUseOnEntityTransaction()
+					{
+
+						EntityId = ReadVarLong(),
+						ActionType = (McpeInventoryTransaction.ItemUseOnEntityAction) ReadVarInt(),
+						Slot = ReadSignedVarInt(),
+						Item = ReadItem(),
+						FromPosition = ReadVector3(),
+						ClickPosition = ReadVector3()
+					};
 					break;
 				case McpeInventoryTransaction.TransactionType.ItemRelease:
-					trans.ActionType = ReadVarInt();
-					trans.Slot = ReadSignedVarInt();
-					trans.Item = ReadItem();
-					trans.FromPosition = ReadVector3();
-					break;
-				default:
+					transaction = new ItemReleaseTransaction()
+					{
+						ActionType = (McpeInventoryTransaction.ItemReleaseAction) ReadVarInt(),
+						Slot = ReadSignedVarInt(),
+						Item = ReadItem(),
+						FromPosition = ReadVector3()
+					};
 					break;
 			}
 
-			return trans;
+			transaction.TransactionRecords = transactions;
+
+			return transaction;
 		}
 
 		public void Write(Item stack)
@@ -965,7 +989,7 @@ namespace MiNET.Net
 			ushort nbtLen = _reader.ReadUInt16(); // NbtLen
 			if (nbtLen == 0xffff && ReadByte() == 1)
 			{
-				stack.ExtraData = ReadNbt().NbtFile.RootTag;
+				stack.ExtraData = (NbtCompound) ReadNbt().NbtFile.RootTag;
 			}
 
 			var canPlace = ReadSignedVarInt();
@@ -1185,24 +1209,56 @@ namespace MiNET.Net
 
 		public BlockPallet ReadBlockPallet()
 		{
-			// don't know how to do it without changes in fNbt
-
 			var result = new BlockPallet();
-			//uint count = ReadUnsignedVarInt();
-			//for (int runtimeId = 0; runtimeId < count; runtimeId++)
-			//{
-			//	var name = ReadString();
-			//	var data = ReadShort();
-			//	var legacyId = ReadShort();
-			//	result.Add(runtimeId, new Blockstate
-			//	{
-			//		Id = legacyId,
-			//		RuntimeId = runtimeId,
-			//		Name = name,
-			//		Data = data
-			//	});
-			//}
 
+			var nbt = new Nbt();
+			var file = new NbtFile();
+			file.BigEndian = false;
+			file.UseVarInt = true;
+			file.AllowAlternativeRootTag = true;
+			nbt.NbtFile = file;
+			file.LoadFromStream(_writer.BaseStream, NbtCompression.None);
+
+			int runtimeId = 0;
+			var rootTag = (NbtList) file.RootTag;
+			foreach (NbtTag tag in rootTag)
+			{
+				var record = new BlockRecord();
+				record.RuntimeId = runtimeId++;
+				record.Id = tag["id"].ShortValue;
+
+				var blockTag = (NbtCompound) tag["block"];
+				record.Name = blockTag["name"].StringValue;
+				if(blockTag.Contains("data")) record.Data = blockTag["data"].ShortValue;
+
+				record.States = new List<BlockState>();
+				var states = (NbtCompound) blockTag["states"];
+				foreach (NbtTag stateTag in states)
+				{
+					var state = new BlockState();
+					state.Name = stateTag.Name;
+					switch (stateTag.TagType)
+					{
+						case NbtTagType.Byte:
+							state.Type = (byte) NbtTagType.Byte;
+							state.Value = stateTag.ByteValue.ToString();
+							break;
+						case NbtTagType.Int:
+							state.Type = (byte) NbtTagType.Int;
+							state.Value = stateTag.IntValue.ToString();
+							break;
+						case NbtTagType.String:
+							state.Type = (byte) NbtTagType.String;
+							state.Value = stateTag.StringValue.ToString();
+							break;
+						default:
+							throw new ArgumentOutOfRangeException();
+					}
+					record.States.Add(state);
+				}
+
+				result.Add(record);
+			}
 			return result;
 		}
 
@@ -1487,65 +1543,78 @@ namespace MiNET.Net
 
 			foreach (Recipe recipe in recipes)
 			{
-				if (recipe is ShapelessRecipe)
+				switch (recipe)
 				{
-					WriteSignedVarInt(Shapeless); // Type
-
-					ShapelessRecipe rec = (ShapelessRecipe) recipe;
-					Write(rec.Id.ToString());
-					WriteVarInt(rec.Input.Count);
-					foreach (Item stack in rec.Input)
+					case ShapelessRecipe shapelessRecipe:
 					{
-						WriteRecipeIngredient(stack);
-					}
-					WriteVarInt(1);
-					Write(rec.Result);
-					Write(rec.Id);
-					Write(rec.Block);
-					WriteSignedVarInt(0); // priority
-				}
-				else if (recipe is ShapedRecipe)
-				{
-					WriteSignedVarInt(Shaped); // Type
+						WriteSignedVarInt(Shapeless); // Type
 
-					ShapedRecipe rec = (ShapedRecipe) recipe;
-					Write(rec.Id.ToString());
-					WriteSignedVarInt(rec.Width);
-					WriteSignedVarInt(rec.Height);
-
-					for (int w = 0; w < rec.Width; w++)
-					{
-						for (int h = 0; h < rec.Height; h++)
+						var rec = shapelessRecipe;
+						Write(rec.Id.ToString());
+						WriteVarInt(rec.Input.Count);
+						foreach (Item stack in rec.Input)
 						{
-							WriteRecipeIngredient(rec.Input[(h * rec.Width) + w]);
+							WriteRecipeIngredient(stack);
 						}
+						WriteVarInt(rec.Result.Count);
+						foreach (Item item in rec.Result)
+						{
+							Write(item);
+						}
+						Write(rec.Id);
+						Write(rec.Block);
+						WriteSignedVarInt(0); // priority
+						break;
 					}
-					WriteVarInt(1);
-					Write(rec.Result);
-					Write(rec.Id);
-					Write(rec.Block);
-					WriteSignedVarInt(0); // priority
-				}
-				else if (recipe is SmeltingRecipe)
-				{
-					SmeltingRecipe rec = (SmeltingRecipe) recipe;
-					WriteSignedVarInt(rec.Input.Metadata == 0 ? Furnace : FurnaceData); // Type
-					WriteSignedVarInt(rec.Input.Id);
-					if (rec.Input.Metadata != 0) WriteSignedVarInt(rec.Input.Metadata);
-					Write(rec.Result);
-					Write(rec.Block);
-				}
-				else if (recipe is MultiRecipe)
-				{
-					WriteSignedVarInt(Multi); // Type
-					Write(recipe.Id);
+					case ShapedRecipe shapedRecipe:
+					{
+						WriteSignedVarInt(Shaped); // Type
+
+						var rec = shapedRecipe;
+						Write(rec.Id.ToString());
+						WriteSignedVarInt(rec.Width);
+						WriteSignedVarInt(rec.Height);
+						for (int w = 0; w < rec.Width; w++)
+						{
+							for (int h = 0; h < rec.Height; h++)
+							{
+								WriteRecipeIngredient(rec.Input[(h * rec.Width) + w]);
+							}
+						}
+						WriteVarInt(rec.Result.Count);
+						foreach (Item item in rec.Result)
+						{
+							Write(item);
+						}
+						Write(rec.Id);
+						Write(rec.Block);
+						WriteSignedVarInt(0); // priority
+						break;
+					}
+					case SmeltingRecipe smeltingRecipe:
+					{
+						var rec = smeltingRecipe;
+						WriteSignedVarInt(rec.Input.Metadata == 0 ? Furnace : FurnaceData); // Type
+						WriteSignedVarInt(rec.Input.Id);
+						if (rec.Input.Metadata != 0)
+						{
+							WriteSignedVarInt(rec.Input.Metadata);
+						}
+						Write(rec.Result);
+						Write(rec.Block);
+						break;
+					}
+					case MultiRecipe _:
+						WriteSignedVarInt(Multi); // Type
+						Write(recipe.Id);
+						break;
 				}
 			}
 		}
 
 		public Recipes ReadRecipes()
 		{
-			Recipes recipes = new Recipes();
+			var recipes = new Recipes();
 
 			int count = (int) ReadUnsignedVarInt();
 
@@ -1563,126 +1632,139 @@ namespace MiNET.Net
 					break;
 				}
 
-				if (recipeType == Shapeless || recipeType == ShulkerBox)
+				switch (recipeType)
 				{
-					ShapelessRecipe recipe = new ShapelessRecipe();
-					ReadString(); // some unique id
-					int ingrediensCount = ReadVarInt(); // 
-					for (int j = 0; j < ingrediensCount; j++)
+					case Shapeless:
+					case ShulkerBox:
 					{
-						recipe.Input.Add(ReadRecipeIngredient());
-					}
-					ReadVarInt(); // 1?
-					recipe.Result = ReadItem();
-					recipe.Id = ReadUUID(); // Id
-					recipe.Block = ReadString(); // block?
-					recipes.Add(recipe);
-					ReadSignedVarInt(); // priority
-					//Log.Error("Read shapeless recipe");
-				}
-				else if (recipeType == Shaped)
-				{
-					ReadString(); // some unique id
-					int width = ReadSignedVarInt(); // Width
-					int height = ReadSignedVarInt(); // Height
-					ShapedRecipe recipe = new ShapedRecipe(width, height);
-					if (width > 3 || height > 3) throw new Exception("Wrong number of ingredience. Width=" + width + ", height=" + height);
-					for (int w = 0; w < width; w++)
-					{
-						for (int h = 0; h < height; h++)
+						var recipe = new ShapelessRecipe();
+						ReadString(); // some unique id
+						int ingrediensCount = ReadVarInt(); // 
+						for (int j = 0; j < ingrediensCount; j++)
 						{
-							recipe.Input[(h * width) + w] = ReadRecipeIngredient();
+							recipe.Input.Add(ReadRecipeIngredient());
 						}
+						int resultCount = ReadVarInt(); // 1?
+						for (int j = 0; j < resultCount; j++)
+						{
+							recipe.Result.Add(ReadItem());
+						}
+						recipe.Id = ReadUUID(); // Id
+						recipe.Block = ReadString(); // block?
+						recipes.Add(recipe);
+						ReadSignedVarInt(); // priority
+						//Log.Error("Read shapeless recipe");
+						break;
 					}
+					case Shaped:
+					{
+						ReadString(); // some unique id
+						int width = ReadSignedVarInt(); // Width
+						int height = ReadSignedVarInt(); // Height
+						var recipe = new ShapedRecipe(width, height);
+						if (width > 3 || height > 3) throw new Exception("Wrong number of ingredience. Width=" + width + ", height=" + height);
+						for (int w = 0; w < width; w++)
+						{
+							for (int h = 0; h < height; h++)
+							{
+								recipe.Input[(h * width) + w] = ReadRecipeIngredient();
+							}
+						}
 
-					int resultCount = ReadVarInt(); // 1?
-					for (int j = 0; j < resultCount; j++)
-					{
-						recipe.Result = ReadItem();
+						int resultCount = ReadVarInt(); // 1?
+						for (int j = 0; j < resultCount; j++)
+						{
+							recipe.Result.Add(ReadItem());
+						}
+						recipe.Id = ReadUUID(); // Id
+						recipe.Block = ReadString(); // block?
+						recipes.Add(recipe);
+						ReadSignedVarInt(); // priority
+						//Log.Error("Read shaped recipe");
+						break;
 					}
-					recipe.Id = ReadUUID(); // Id
-					recipe.Block = ReadString(); // block?
-					recipes.Add(recipe);
-					ReadSignedVarInt(); // priority
-					//Log.Error("Read shaped recipe");
-				}
-				else if (recipeType == Furnace)
-				{
-					SmeltingRecipe recipe = new SmeltingRecipe();
-					//short meta = (short) ReadVarInt(); // input (with metadata) 
-					short id = (short) ReadSignedVarInt(); // input (with metadata) 
-					Item result = ReadItem(); // Result
-					recipe.Block = ReadString(); // block?
-					recipe.Input = ItemFactory.GetItem(id, 0);
-					recipe.Result = result;
-					recipes.Add(recipe);
-					//Log.Error("Read furnace recipe");
-					//Log.Error($"Input={id}, meta={""} Item={result.Id}, Meta={result.Metadata}");
-				}
-				else if (recipeType == FurnaceData)
-				{
-					//const ENTRY_FURNACE_DATA = 3;
-					SmeltingRecipe recipe = new SmeltingRecipe();
-					short id = (short) ReadSignedVarInt(); // input (with metadata) 
-					short meta = (short) ReadSignedVarInt(); // input (with metadata) 
-					Item result = ReadItem(); // Result
-					recipe.Block = ReadString(); // block?
-					recipe.Input = ItemFactory.GetItem(id, meta);
-					recipe.Result = result;
-					recipes.Add(recipe);
-					//Log.Error("Read smelting recipe");
-					//Log.Error($"Input={id}, meta={meta} Item={result.Id}, Meta={result.Metadata}");
-				}
-				else if (recipeType == Multi)
-				{
-					//Log.Error("Reading MULTI");
+					case Furnace:
+					{
+						var recipe = new SmeltingRecipe();
+						//short meta = (short) ReadVarInt(); // input (with metadata) 
+						short id = (short) ReadSignedVarInt(); // input (with metadata) 
+						Item result = ReadItem(); // Result
+						recipe.Block = ReadString(); // block?
+						recipe.Input = ItemFactory.GetItem(id, 0);
+						recipe.Result = result;
+						recipes.Add(recipe);
+						//Log.Error("Read furnace recipe");
+						//Log.Error($"Input={id}, meta={""} Item={result.Id}, Meta={result.Metadata}");
+						break;
+					}
+					case FurnaceData:
+					{
+						//const ENTRY_FURNACE_DATA = 3;
+						var recipe = new SmeltingRecipe();
+						short id = (short) ReadSignedVarInt(); // input (with metadata) 
+						short meta = (short) ReadSignedVarInt(); // input (with metadata) 
+						Item result = ReadItem(); // Result
+						recipe.Block = ReadString(); // block?
+						recipe.Input = ItemFactory.GetItem(id, meta);
+						recipe.Result = result;
+						recipes.Add(recipe);
+						//Log.Error("Read smelting recipe");
+						//Log.Error($"Input={id}, meta={meta} Item={result.Id}, Meta={result.Metadata}");
+						break;
+					}
+					case Multi:
+					{
+						//Log.Error("Reading MULTI");
 
-					MultiRecipe recipe = new MultiRecipe();
-					recipe.Id = ReadUUID();
-					recipes.Add(recipe);
-				}
-				else if (recipeType == ShapelessChemistry)
-				{
-					int ingrediensCount = ReadVarInt(); // 
-					for (int j = 0; j < ingrediensCount; j++)
-					{
-						ReadItem();
+						var recipe = new MultiRecipe();
+						recipe.Id = ReadUUID();
+						recipes.Add(recipe);
+						break;
 					}
-					int resultCount = ReadVarInt(); // 
-					for (int j = 0; j < resultCount; j++)
+					case ShapelessChemistry:
 					{
-						ReadItem();
-					}
-
-					ReadUUID();
-					ReadString(); // block?
-				}
-				else if (recipeType == ShapedChemistry)
-				{
-					int width = ReadSignedVarInt(); // Width
-					int height = ReadSignedVarInt(); // Height
-					if (width > 3 || height > 3) throw new Exception("Wrong number of ingredience. Width=" + width + ", height=" + height);
-					for (int w = 0; w < width; w++)
-					{
-						for (int h = 0; h < height; h++)
+						int ingrediensCount = ReadVarInt(); // 
+						for (int j = 0; j < ingrediensCount; j++)
 						{
 							ReadItem();
 						}
-					}
+						int resultCount = ReadVarInt(); // 
+						for (int j = 0; j < resultCount; j++)
+						{
+							ReadItem();
+						}
 
-					int resultCount = ReadVarInt(); // 1?
-					for (int j = 0; j < resultCount; j++)
+						ReadUUID();
+						ReadString(); // block?
+						break;
+					}
+					case ShapedChemistry:
 					{
-						ReadItem();
-					}
+						int width = ReadSignedVarInt(); // Width
+						int height = ReadSignedVarInt(); // Height
+						if (width > 3 || height > 3) throw new Exception("Wrong number of ingredience. Width=" + width + ", height=" + height);
+						for (int w = 0; w < width; w++)
+						{
+							for (int h = 0; h < height; h++)
+							{
+								ReadItem();
+							}
+						}
 
-					ReadUUID(); // Id
-					ReadString(); // block?
-				}
-				else
-				{
-					Log.Error($"Read unknown recipe type: {recipeType}");
-					//ReadBytes(len);
+						int resultCount = ReadVarInt(); // 1?
+						for (int j = 0; j < resultCount; j++)
+						{
+							ReadItem();
+						}
+
+						ReadUUID(); // Id
+						ReadString(); // block?
+						break;
+					}
+					default:
+						Log.Error($"Read unknown recipe type: {recipeType}");
+						//ReadBytes(len);
+						break;
 				}
 			}
 
@@ -1714,6 +1796,51 @@ namespace MiNET.Net
 			int count = ReadSignedVarInt();
 
 			return ItemFactory.GetItem(id, metadata, count);
+		}
+
+
+		public void Write(PotionContainerChangeRecipe[] recipes)
+		{
+
+		}
+
+		public PotionContainerChangeRecipe[] ReadPotionContainerChangeRecipes()
+		{
+			int count = (int) ReadUnsignedVarInt();
+			var recipes = new PotionContainerChangeRecipe[count];
+			for (int i = 0; i < recipes.Length; i++)
+			{
+				var recipe = new PotionContainerChangeRecipe();
+				recipe.InputItemId = ReadVarInt();
+				recipe.IngredientItemId = ReadVarInt();
+				recipe.OutputItemId = ReadVarInt();
+
+				recipes[i] = recipe;
+			}
+
+			return recipes;
+		}
+
+		public void Write(PotionTypeRecipe[] recipes)
+		{
+
+		}
+
+		public PotionTypeRecipe[] ReadPotionTypeRecipes()
+		{
+			int count = (int) ReadUnsignedVarInt();
+			var recipes = new PotionTypeRecipe[count];
+			for (int i = 0; i < recipes.Length; i++)
+			{
+				var recipe = new PotionTypeRecipe();
+				recipe.InputPotionType = ReadVarInt();
+				recipe.IngredientItemId = ReadVarInt();
+				recipe.OutputPotionType = ReadVarInt();
+
+				recipes[i] = recipe;
+			}
+
+			return recipes;
 		}
 
 		const int BITFLAG_TEXTURE_UPDATE = 0x02;
