@@ -1,4 +1,4 @@
-#region LICENSE
+﻿#region LICENSE
 
 // The contents of this file are subject to the Common Public Attribution
 // License Version 1.0. (the "License"); you may not use this file except in
@@ -52,6 +52,7 @@ namespace MiNET.Net
 		StartSwimming = 21,
 		StopSwimming = 22,
 		StartSpinAttack = 23,
-		StopSpinAttack = 24
+		StopSpinAttack = 24,
+		InteractBlock = 25
 	}
 }
