@@ -383,15 +383,19 @@ namespace MiNET
 
 				MiNetServer.TraceReceive(message);
 
-				if (CryptoContext != null && CryptoContext.UseEncryption)
-				{
-					MiNetServer.FastThreadPool.QueueUserWorkItem(() =>
-					{
-						HandlePacket(MessageHandler, message as Packet);
-						message.PutPool();
-					});
-				}
-				else
+				// This needs rework. Some packets (transaction) really require the correct order to be maintained 
+				// in order to process correctly. I was thinking perhaps marking some packets as "order important" and
+				// check that flag here. But for now, I'm disabling the threading here. Not even sure why it is here.
+
+				//if (CryptoContext != null && CryptoContext.UseEncryption)
+				//{
+				//	MiNetServer.FastThreadPool.QueueUserWorkItem(() =>
+				//	{
+				//		HandlePacket(MessageHandler, message as Packet);
+				//		message.PutPool();
+				//	});
+				//}
+				//else
 				{
 					HandlePacket(MessageHandler, message);
 					message.PutPool();
@@ -473,6 +477,11 @@ namespace MiNET
 			else if (typeof(McpeLevelSoundEvent) == message.GetType())
 			{
 				handler.HandleMcpeLevelSoundEvent((McpeLevelSoundEvent) message);
+			}
+
+			else if (typeof(McpeClientCacheStatus) == message.GetType())
+			{
+				handler.HandleMcpeClientCacheStatus((McpeClientCacheStatus) message);
 			}
 
 			else if (typeof(McpeAnimate) == message.GetType())
@@ -653,6 +662,11 @@ namespace MiNET
 			else if (typeof(McpeSetEntityData) == message.GetType())
 			{
 				handler.HandleMcpeSetEntityData((McpeSetEntityData) message);
+			}
+
+			else if (typeof(McpeTickSync) == message.GetType())
+			{
+				handler.HandleMcpeTickSync((McpeTickSync) message);
 			}
 
 			else if (typeof(McpeNpcRequest) == message.GetType())
